requirements.txt
pyyaml
requests
retry
